+++ CHANGELOG.md
@@ -17,6 +17,15 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Breaks
 
 
+## 0.2.1 - (2022-02-10)
+---
+
+### New
+* services.cfn.deploy_template supports input parameters
+* Enable toggling aborting CodeBuild phases on command execution failure
+* Support CodeBuild spec runtime-versions
+
+
 ## 0.2.0 - (2022-02-03)
 ---